Delivery Service Done

+++ src/main/java/org/kainos/ea/cli/DeliveryRequest.java
@@ -17,11 +17,11 @@ public class DeliveryRequest {
         this.name = name;
     }
 
-    public double getSalary() {
+    public float getSalary() {
         return salary;
     }
 
-    public void setSalary(double salary) {
+    public void setSalary(float salary) {
         this.salary = salary;
     }
